Maude Alpha95b
Dear Maude Abusers,

This version identifies itself as Maude 2.6 and will be released as
such unless a show stopper is reported.

Bug fixes
-----------
A bug in the profiler. Essentially the profiler is called each time a
condition fragment is executed, but if the condition fragment does not
belong to a statement in the module the profiler doesn't know what to
do with it. And condition fragments with this property can in fact
belong to a command such as search or an invocation of a descent
function such as metaSearch(). Bug provoked by an example from
Francesco Spegni.

The ideal solution would be to allow profiling of these fragments, but
it's not so simple since multiple commands (by interrupting one
command and starting another) or multiple descent function invocations
might be in play at the same time and the profiler would need to keep
track of all of them rather than just a static module. The profiler is
currently implemented as a module subclass and doesn't know about
commands or the metalevel.

Instead I've done the same thing as already happens for tracing -
condition fragments not belong to a module statement are detected and
not profiled. Of course the execution of statements in the evaluation
of such a condition fragment will still be profiled as usual.

Other changes
--------------
(1) Only left identity sort checks are now performed for operators
which have both the comm and id: attributes since right identity sort
checks can only produce redundant warnings in this case.

(2) More overparsing - gracefully recovers from syntax errors of the
form
  vars X Y : -> Elt .
in modules/theories/views and of the form:
  op f : Foo Bar Foo : -> Bar .
in modules and theories.

Currently there is only a Linux64 binary. Our IntelDarwin server is
down and our PPC/Darwin server has been disposed of (Apple no longer
supports this hardware). Does anyone still want a staticallly linked
32-bit Linux version or can this be dropped for the new release?

Steven