Maude Alpha95c
Dear Maude Abusers,

This version identifies itself as Maude 2.7 and will be released as
such unless a show stopper is reported. Currently there is only a
Linux64 binary - the 32-bit Linux and Darwin servers seem to offline
for the vacation. Users of those systems can always build from source.

This is an end of year bug fix release.

Bug fixes
-----------
(1) A bug in the handling of fresh variables introduced by ACU
unification that led to some internal consistancy checking failing in
the debug build of Maude. Reported by Kyungmin Bae.

(2) A bug in the handling of writes to a socket that could potentially
access memory that has been deallocated and write garbage characters
to the socket - discovered while searching for (3).

(3) A bug that appears to be in the rope class template; I have avoided
using the buggy function at some performance cost. This was causing
what appeared to be a socket bug that was reported by Traian.

Steven