Maude Alpha96b
Dear Maude Abusers,

Bug fixes
--------

(1) A bug in variant  narrowing where trying to generate variants of a ground term resulted in a crash. Reported both by Santiago and Kyungmin Bae.

(2) A bug which caused memory corruption with sockets; this was reported by Grigore Rosu on 64-bit machines, but it occurs with less visibility on all machines.

(3) A long standing bug in sockets where sockets appear to be dropped on Macs. This was actually reported by Musab AlTurki <musab.alturki@gmail.com> back
on May 11,  2009 and actually results from the interaction of two separate bugs and the MacOS implementation of POLLHUP under certain race conditions. I originally
thought this might be the same bug as (2), hence the renewed effort to find it. It was originally reported for 32-bit Macs but would affect any *NIX operating system
that reports both POLLIN and POLLHUP when a process is checking for readable data on a socket with poll() and the other end of the socket has be closed. Linux only reports POLLIN which avoids triggering the bugs and the potential race condition.

Steven