Maude Alpha100
Dear Maude Abusers,

This version is more alpha-quality than usual - it's sole raison d'etre it to allow benchmarking of
a new explicit stack based interpreter - the Maude Virtual Machine (MVM).

The currently implementation is minimalist - it's really just a proof of concept - although it's
designed to be extensible so in principle it could be extended to do everything the current interpreter
does, and potentially things that are impossible with the current design.

Only the free theory is supported, and equations must be left linear.
There are no built-ins (not even BOOL), and conditional equations are not supported.
There is no tracing, profiling, break points, or ^C interrupts. Stack overflow produces a
crash (just as with the regular interpreter) but it is not identified as a stack overflow.

To use the MVM, instead of the usual interpreter, use sred in place of red. There are some
benchmarks in tests/Misc/sreduce

The main advantages of the MVM approach are:
(1) Dag nodes on only constructed when reduced (saves time).
(2) No in-place modification of dags (which plays havoc with things like variant narrowing and
requires dag copying for safety with regular interpreter).
(3) Entire computation is a C++ object that in principle could be interrupted, copied
potentially even serialized to a file, then restarted at a later date.
(4) Avoids the system stack limits on Mac.
(5) Potential to handle stack overflows gracefully in the future.
(6) The MVM is about 50% faster than the current interpreter for the Maude fragment
that it handles (which is also the fragment that runs fastest on the current interpreter).

In case anyone is wondering why this is Alpha100 when the last alpha release was Alpha96c,
I number the alpha releases by source code tree and I abandoned trees 97, 98 and 99 - making
a new interpreter that consistently beats the current one while sharing dag representation and
memory management turned out to be harder than expected.

Steven