Maude Alpha100a
Dear Maude Abusers,

This version reports itself as Maude 2.7 and will be released as such if no show stoppers are
found.

Changes
=======
(1) A bug reported by Antonio Burgueno where debugging messages were generated
during variant generation is fixed.

(2) The restriction on non-linear left-hand-sides in equations used by sreduce is lifted.

(3) The Mac version is now compiled with the Mac Ports toolchain rather than the official
Apple one. This resolved a number of issues; in particular:
(a) A bug where Maude would hang if it hit the end-of-file while processing a comment
(reported by Carolyn almost a year ago). This was caused by a bug in the Apple version of
flex where an end-of-file was being returned as an infinite stream of NULLs.
(b) Problems compiling the rope extension with the Apple version of gcc (which isn't gcc
at all but an ancient (c. 2007) front end to llvm).

(4) Much code cleaning to eliminate most compiler warnings (remaining warnings are
benign and difficult to avoid). Doug Lea's malloc has been removed from 3rdParty as it
causes issues on various platforms and it is no longer needed for performance.

Steven