Maude Alpha101
Dear Maude Abusers,

This version reports itself as Maude 2.7 and will be released as such if no show stoppers are
found.

Changes
=======
(1) A bug where debugging messages were generated during rule rewriting on very deep terms
(reported by Carolyn).

(2) sreduce now supports _==_ and _=/=_

(3) Variant narrowing and unification are now partly incrementalized - this means that rather
than compute the complete set of variants up front, Maude now computes them layer at
a time as needed. It's not fully incremental in that a full layer of narrowing steps is computed
in one go, and this shows up in the rewrite counts (which are now output for each variant).

One downside to returning variants before they have all been generated is that it is possible
that a variant might be returned which is later found to be redundant. This isn't really an
issue for variant unification, since only unifiers are returned, and variant unification does
not guarantee a minimal complete set of unifiers anyway. However for
variant generation, the old semantics might be preferable for some applications, and
is therefore available with the command:

get irredundant variants <term> .

which has symmetric extended syntax with the get variants command. This is reflected
by the descent function:

  op metaGetIrredundantVariant : Module Term TermList Nat Nat ~> Variant? [special (...)] .

(4) There is a new command line option -always-advise which disables the possibility of
turning advisories off. This is mostly for my use, since I end up running other peoples
Maude programs that turn advisories off and miss valuable debugging information.

Steven